feat: move authenticator objects to `auth` module (#714)
