+++ miden-tx/src/auth/mod.rs
@@ -1,0 +1,4 @@
+mod tx_authenticator;
+pub use tx_authenticator::{BasicAuthenticator, TransactionAuthenticator};
+
+pub mod signatures;
